Replaced uint32 in OpenGLES2

+++ src/renderers/OpenGLES2/opengles2shape3dinterface.go
@@ -13,7 +13,7 @@ type OpenGLES2Shape3DInterface struct {
 
 	points      []gohome.Shape3DVertex
 	drawMode    uint32
-	numVertices uint32
+	numVertices int
 	lineWidth   float32
 }
 
@@ -68,7 +68,7 @@ func (this *OpenGLES2Shape3DInterface) Load() {
 		return
 	}
 
-	this.numVertices = uint32(len(this.points))
+	this.numVertices = len(this.points)
 	if this.numVertices == 0 {
 		gohome.ErrorMgr.Error("Shape3DInterface", this.Name, "No Vertices have been added!")
 		return
